config: make Config finalizer panic if Close() isn't explicitly deferred/called
When using a finalizer to free the config object C pointer, it often
appears to overzealously run the finalizer during a `for` loop within
its scope.

Forcing use of explicit Close() (or deferred Close()) makes it so the
config object is visible to the GC throughout its entire scope lifetime,
avoiding freeing it when entering loop sub-scopes or similar.